Moved get_reduction_function() to models/torch_modules/__init__.py
