+++ dna/utils.py
@@ -12,14 +12,3 @@ def get_git_commit_hash():
         return git.Repo(search_parent_directories=True).head.object.hexsha
     except git.exc.InvalidGitRepositoryError:
         pass
-
-def get_reduction_function(reduction: str):
-    if reduction == 'mean':
-        return torch.mean
-    elif reduction == 'sum':
-        return torch.sum
-    elif reduction == 'mul':
-        return torch.mul
-    else:
-        raise Exception('No valid reduction was provided\n'
-                        'Got \"' + reduction + '\"')
